// packages/backend/src/types/schemas/prisma-generated/inputTypeSchemas/PendingSubscriptionUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { PlanTypeSchema } from './PlanTypeSchema';

export const PendingSubscriptionUncheckedCreateInputSchema: z.ZodType<Prisma.PendingSubscriptionUncheckedCreateInput> = z.object({
  id: z.string().optional(),
  userId: z.string(),
  planType: z.lazy(() => PlanTypeSchema),
  billingCycle: z.string(),
  status: z.string().optional(),
  createdAt: z.coerce.date().optional(),
  updatedAt: z.coerce.date().optional()
}).strict();

export default PendingSubscriptionUncheckedCreateInputSchema;